requirements.txt
anthropic>=0.43.0
flask>=3.1.0
requests>=2.32.0
python-dotenv>=1.0.1
gunicorn>=23.0.0
